refactor: organize memorization modal subcomponents
