+++ frontend/src/app/shared/components/memorization-modal/components/completion-stage/completion-stage.component.html
