+++ frontend/src/app/shared/components/memorization-modal/components/confirmation-modal/confirmation-modal.component.ts
