+++ frontend/src/app/shared/components/memorization-modal/components/navigation-controls/navigation-controls.component.html
