+++ frontend/src/app/shared/components/memorization-modal/components/practice-stage/practice-stage.component.html
