+++ frontend/src/app/shared/components/memorization-modal/components/progress-journey/progress-journey.component.ts
