+++ frontend/src/app/shared/components/memorization-modal/components/settings-menu/settings-menu.component.html
