+++ frontend/src/app/shared/components/memorization-modal/components/setup-stage/setup-stage.component.ts
