+++ frontend/src/app/shared/components/memorization-modal/components/verse-bubbles/verse-bubbles.component.ts
