+++ frontend/src/app/shared/components/memorization-modal/memorization-modal.component.ts
@@ -17,14 +17,14 @@ import { Subject, takeUntil } from 'rxjs';
 import { trigger, style, transition, animate, keyframes, state } from '@angular/animations';
 
 // Import all child components
-import { SetupStageComponent } from './setup-stage/setup-stage.component';
-import { ProgressJourneyComponent, ProgressMarker } from './progress-journey/progress-journey.component';
-import { VerseBubblesComponent, Verse } from './verse-bubbles/verse-bubbles.component';
-import { PracticeStageComponent, PracticeSettings } from './practice-stage/practice-stage.component';
-import { CompletionStageComponent } from './completion-stage/completion-stage.component';
-import { SettingsMenuComponent } from './settings-menu/settings-menu.component';
-import { NavigationControlsComponent } from './navigation-controls/navigation-controls.component';
-import { ConfirmationModalComponent } from './confirmation-modal/confirmation-modal.component';
+import { SetupStageComponent } from './components/setup-stage/setup-stage.component';
+import { ProgressJourneyComponent, ProgressMarker } from './components/progress-journey/progress-journey.component';
+import { VerseBubblesComponent, Verse } from './components/verse-bubbles/verse-bubbles.component';
+import { PracticeStageComponent, PracticeSettings } from './components/practice-stage/practice-stage.component';
+import { CompletionStageComponent } from './components/completion-stage/completion-stage.component';
+import { SettingsMenuComponent } from './components/settings-menu/settings-menu.component';
+import { NavigationControlsComponent } from './components/navigation-controls/navigation-controls.component';
+import { ConfirmationModalComponent } from './components/confirmation-modal/confirmation-modal.component';
 
 // Import types
 import { ReviewStage, StarPopup, AnimatedStar } from './memorization-modal.types';
